+++ app/src/main/AndroidManifest.xml
@@ -1,16 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.vimteam.lesson1">
+    package="org.vimteam.weatherreport">
 
     <application
-        android:name=".App"
+        android:name="org.vimteam.weatherreport.App"
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@mipmap/ic_weather_report"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:roundIcon="@mipmap/ic_weather_report"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Lesson1">
-        <activity android:name=".main.ui.activities.MainActivity">
+        android:theme="@style/Theme.WeatherReport">
+        <activity
+            android:name="org.vimteam.weatherreport.main.ui.activities.MainActivity">
+
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
