--- app/src/main/java/org/vimteam/lesson1/main/base/MainConstants.kt
@@ -1,6 +1,0 @@
-package org.vimteam.lesson1.main.base
-
-object MainConstants {
-
-    const val LOG_TAG = "L1"
-}
