--- app/src/main/java/org/vimteam/lesson1/main/domain/contracts/WeatherRepositoryContract.kt
@@ -1,8 +1,0 @@
-package org.vimteam.lesson1.main.domain.contracts
-
-import org.vimteam.lesson1.main.domain.models.WeatherData
-
-interface WeatherRepositoryContract {
-
-    fun getStoredData(): WeatherData
-}
