--- app/src/main/java/org/vimteam/lesson1/main/domain/models/WeatherData.kt
@@ -1,8 +1,0 @@
-package org.vimteam.lesson1.main.domain.models
-
-data class WeatherData(
-        val city: String,
-        val temperature: Float,
-        val wind: Wind
-
-)
