--- app/src/main/java/org/vimteam/lesson1/main/domain/models/Wind.kt
@@ -1,6 +1,0 @@
-package org.vimteam.lesson1.main.domain.models
-
-data class Wind(
-        val speed: Int,
-        val direction: WindDirection
-)
