--- app/src/main/java/org/vimteam/lesson1/main/domain/models/WindDirection.kt
@@ -1,8 +1,0 @@
-package org.vimteam.lesson1.main.domain.models
-
-enum class WindDirection {
-    NORTH,
-    SOUTH,
-    WEST,
-    EAST
-}
