+++ app/src/main/java/org/vimteam/weatherreport/main/base/MainConstants.kt
@@ -1,0 +1,6 @@
+package org.vimteam.weatherreport.main.base
+
+object MainConstants {
+
+    const val LOG_TAG = "L1"
+}
