+++ app/src/main/java/org/vimteam/weatherreport/main/domain/contracts/WeatherRepositoryContract.kt
@@ -1,0 +1,8 @@
+package org.vimteam.weatherreport.main.domain.contracts
+
+import org.vimteam.weatherreport.main.domain.models.WeatherData
+
+interface WeatherRepositoryContract {
+
+    fun getStoredData(): WeatherData
+}
