+++ app/src/main/java/org/vimteam/weatherreport/main/domain/models/WeatherData.kt
@@ -1,0 +1,8 @@
+package org.vimteam.weatherreport.main.domain.models
+
+data class WeatherData(
+        val city: String,
+        val temperature: Float,
+        val wind: Wind
+
+)
