+++ app/src/main/java/org/vimteam/weatherreport/main/domain/models/Wind.kt
@@ -1,0 +1,6 @@
+package org.vimteam.weatherreport.main.domain.models
+
+data class Wind(
+        val speed: Int,
+        val direction: WindDirection
+)
