+++ app/src/main/java/org/vimteam/weatherreport/main/domain/models/WindDirection.kt
@@ -1,0 +1,8 @@
+package org.vimteam.weatherreport.main.domain.models
+
+enum class WindDirection {
+    NORTH,
+    SOUTH,
+    WEST,
+    EAST
+}
