+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name" translatable="false">Lesson1</string>
+    <string name="app_name" translatable="false">Weather Report</string>
     <string name="load">Load</string>
     <string name="day_night">Day/Night</string>
     <string name="select_date">Select date</string>
